requirements.txt
annotated-types==0.7.0
boto3==1.34.122
botocore==1.34.122
dnspython==2.6.1
jmespath==1.0.1
motor==3.4.0
odmantic==1.0.2
pydantic==2.7.3
pydantic_core==2.18.4
pymongo==4.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.1
six==1.16.0
typing_extensions==4.12.1
urllib3==2.2.1
